Refactoring: phase 1
The general trend of this refactoring is to the reduce the proliferation of parametric types which were stifling the previous version. Basically, here we use type classes wherever possible, as opposed to hierarchical types.
Too many changes to detail but Evolvable has essentially been split into Evolvable and Evolute. Colony has been temporarily changed to OldColony and the new Colony will extend Evolute, not Evolvable.